LayoutTests/http/tests/media/media-source/video-media-source-errors-expected.txt

Tests error cases with MediaSource API


running didNotSendInitSegmentFirst
EVENT(loadstart)
EVENT(webkitsourceopen)
Test appending a media segment before sending the initialization segment.
EVENT(error) : MEDIA_ERR_SRC_NOT_SUPPORTED

running immediateDecodeErrorViaEndOfStream
EVENT(emptied)
EVENT(loadstart)
EVENT(webkitsourceopen)
Test signalling a decode error with webkitSourceEndOfStream.
EVENT(webkitsourceended)
EVENT(error) : MEDIA_ERR_SRC_NOT_SUPPORTED

running immediateNetworkErrorViaEndOfStream
EVENT(emptied)
EVENT(loadstart)
EVENT(webkitsourceopen)
Test signalling a network error with webkitSourceEndOfStream.
EVENT(webkitsourceended)
EVENT(error) : MEDIA_ERR_SRC_NOT_SUPPORTED

running decodeErrorAfterHaveMetadata
EVENT(emptied)
EVENT(loadstart)
EVENT(webkitsourceopen)
Test signalling a decode error with webkitSourceEndOfStream after we have appended enough data to get to the HAVE_METADATA state.
EVENT(loadedmetadata)
EVENT(webkitsourceended)
EVENT(error) : MEDIA_ERR_DECODE

running networkErrorAfterHaveMetadata
EVENT(emptied)
EVENT(loadstart)
EVENT(webkitsourceopen)
Test signalling a network error with webkitSourceEndOfStream after we have appended enough data to get to the HAVE_METADATA state.
EVENT(loadedmetadata)
EVENT(webkitsourceended)
EVENT(error) : MEDIA_ERR_NETWORK
END OF TEST
